Terminado simulacion de Jornada. Añadido método para el cargar Datos y clase CargarDatos. Falta mover el código a CargarDatos y dejar el main limpio

+++ src/main/java/entidades/Jugador.java
@@ -19,7 +19,7 @@ public class Jugador {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "IdPlayer")
-	private Long idDeportista;
+	private int idDeportista;
 
 	@Column(name = "License")
 	private int licencia;
@@ -100,7 +100,7 @@ public class Jugador {
 		this.equipo = equipo;
 	}
 
-	public Long getIdDeportista() {
+	public int getIdDeportista() {
 		return idDeportista;
 	}
